+++ app/root.tsx
@@ -31,7 +31,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
         <Meta />
         <Links />
       </head>
-      <body>
+      <body className="dark:bg-gray-900">
         {children}
         <ScrollRestoration />
         <Scripts />